requirements.txt
asgiref==3.8.1
Django==5.0.4
django-environ==0.11.2
django-js-asset==2.2.0
django-mptt==0.16.0
sqlparse==0.4.4
